pyproject.toml
[project]
name = "nexus-agent"
version = "0.1.0"
description = "Agentic AI workflow for Project Nexus (Healthcare) with Zoho integration"
authors = [{ name = "Agentic AI" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "tenacity>=9.0",
  "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "pre-commit>=3.7",
  "detect-secrets>=1.5",
]

[project.scripts]
nexus-agent = "nexus_agent.agent_loop:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "SIM"]

[tool.pytest.ini_options]
addopts = "-q --cov=nexus_agent --cov-report=term-missing"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
